DO NOT MERGE ME YET. THINGS ARE BROKEN
Added a few changes to the booking form, need to fix up the customer input so the phone number displays correctly.

Also made some edits to the CSS, trying to get the navbar to be more like the one shown in the bootstrap example (Fixed Posistion, collapses until scroll up.) . (Might see if I can get the list actions in the collapsable menu as well).